renamed several files and moved localconnectorinterface to separate file
