+++ src/network/defaultlocalconnector.cpp
@@ -1,4 +1,4 @@
-#include "localconnector.h"
+#include "defaultlocalconnector.h"
 // ==========================================================================================
 // ==========================================================================================
 //                                                             LAUNCH SERVICES ON USER LOG IN
