+++ src/network/network.pri
@@ -2,14 +2,15 @@ DEPENDPATH += $$PWD
 INCLUDEPATH += $$PWD
 
 SOURCES += \
-        $$PWD/localconnector.cpp \
+        $$PWD/defaultlocalconnector.cpp \
         $$PWD/localconnectorgenerator.cpp \
         $$PWD/tcpservice.cpp \
         $$PWD/udpservice.cpp
 
 HEADERS += \
-    $$PWD/localconnector.h \
+    $$PWD/defaultlocalconnector.h \
     $$PWD/localconnectorgenerator.h \
+    $$PWD/localconnectorinterface.h \
     $$PWD/networkbase.h \
     $$PWD/tcpservice.h \
     $$PWD/udpservice.h
